server/requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
sqlalchemy
httpx
PyJWT
pwdlib[argon2]
pydantic
pydantic[email]
